requirements.txt
Flask==3.1.3
numpy==2.4.4
opencv-python-headless==4.13.0.92
gunicorn==21.2.0
psycopg2-binary
werkzeug